python cloud functions/dataAnalysisWorkflow/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.38.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
cloudevents==1.2.0
colorama==0.4.4
deprecation==2.1.0
dialogflow==1.1.0
firebase-admin==5.0.0
Flask==2.0.1
functions-framework==2.2.1
google-api-core==1.28.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
google-cloud-core==1.6.0
google-cloud-dialogflow==2.4.0
google-cloud-firestore==2.1.1
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3
pandas==1.0.5
openpyxl==3.0.7